- remove most warnings, fix few ui constraints and clean code up a little

+++ GIF Keyboard/MMKeyboardCollectionViewFlowLayout.m
@@ -36,8 +36,8 @@
 			continue;
 		}
 
-		if (fabsf(attributes.center.x - proposedContentOffsetCenterX) <
-				fabsf(candidateAttributes.center.x - proposedContentOffsetCenterX)) {
+		if (fabs(attributes.center.x - proposedContentOffsetCenterX) <
+				fabs(candidateAttributes.center.x - proposedContentOffsetCenterX)) {
 			candidateAttributes = attributes;
 		}
 	}
